requirements.txt
fastapi
uvicorn
streamlit
requests
groq
pydantic
transformers
torch
python-dotenv